+++ ChangeLog.md
@@ -9,6 +9,7 @@ Potentially breaking changes are marked with an exclamation point '!' at the beg
  * ! Removed all support for stored IndexedTensor*s.
  * ! Reworked the IndexedTensor* index assignment.
  * ! Changed resize_dimension to allow slate insertion at the first position.
+ * Added TTTensor::random with callback function to manipulate the singular values of all matricisations.
  * Support for low level factorisation calls for Tensor class.
  * Several bug fixes, including SVD factor handling, SparseTensor summation, Tensor resize_dimension.
  * Added several test cases.